refactor: use snake_case naming
