+++ README.md
@@ -1,0 +1,15 @@
+# Sofia Memory Plugin
+
+Проект предоставляет API для работы с файлами памяти и планами обучения.
+
+## Кодстайл
+
+Весь код следует одному из трёх стилей именования:
+
+- `snake_case`
+- `UPPER_SNAKE_CASE`
+- `И_Мой_Стиль`
+
+Иные стили (camelCase, PascalCase и т.д.) не допускаются.
+
+Перед отправкой изменений убедитесь, что новые имена и файлы соответствуют данному соглашению.
