--- memoryWithIndex.js
@@ -1,5 +1,0 @@
-const { saveMemoryWithIndex } = require('./core/storage');
-
-module.exports = {
-  saveMemoryWithIndex,
-};
